Significant changes to enable reuse of data structures in LU decomposition linear solver.
All linear solvers have one single setMatrix method from the base class. This method now calls the initializeDataStructure for each derived class that initializes the relevant data structures. This is specifically for sparse matrix LU decomposition, which requires an expensive union of two data structures, but the method is intended to be general for all matrix types.

Nonlinear solver, transient solver and eigensolvers have been modified to enable this feature, and if the user intends to enable reuse of data structure, the he/she should tell the method in each of these nonlinear, transient and eigensolvers to enable this flag for the solver.

Another significant change is the removal of the dirichlet boundary conditions in the Euler solver. This is to save the excessive CPU cost of getting sub matrices and sub vectors for each time step.

More test cases to be run, for evaluation of the updates.