Added step control to Newton nonlinear solver. This creates a second order approximation of the residual and uses that to get to reduce the step size if the residual increases at the full newton step.
Multiple bug fixes and enhancements for the elliptic meshing routine. Most importantly, a coefficient is now used for the coupling terms (g12). These are set to zero close to the curved boundary to obtain close to normal meshing, and then ramped up away from this boundary to try to get more well behaved meshing.